% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrasterval.R
\name{addrasterval}
\alias{addrasterval}
\title{addrasterval}
\usage{
addrasterval(x, r, par = TRUE, dec = 0, rvalname = "Cota")
}
\arguments{
\item{x}{sf layer of points or polygons (not yet tested with lines)}

\item{r}{raster layer (must be in the same CRS as layer)}

\item{par}{if TRUE --> parallel computation}

\item{dec}{number of decimals of the value extracted}

\item{rvalname}{name of the attribute with the raster values}
}
\description{
the function extract raster value for sf layer. If the layer is 
a layer of points --> extract values, if it is a polygon layer it extract the mean value
}
\author{
cesarkero
}
