% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayfootprint.R
\name{dayfootprint}
\alias{dayfootprint}
\title{dayfootprint}
\usage{
dayfootprint(
  layer,
  day = "2021-12-21",
  H = "Altura",
  outdir = "./outputs/",
  tz = "Europe/Paris",
  Xmin = 30,
  stylefile = "./styles/footprint_azul.qml"
)
}
\arguments{
\item{layer}{geographic layer with the obstacles to calculate footprint}

\item{day}{time to start calculation}

\item{H}{attribute of the layer that contains the height of the obstacles}

\item{outdir}{output folder. A subfolder will be created}

\item{tz}{time zone ("Europe/Paris" by default)}

\item{Xmin}{number of minutes between each calculation}

\item{stylefile}{path to the style file asociated to the results (not mandatory)}
}
\description{
This function craetes a single gpkg with the footprints of 
the sun over a layer of buildings (with heigth attribute). The footprints will
be calculated from a vector of date-times, created internally. This vector is
based in the start time (t1) and stop time (t2) and takes into consideration
the possible time restrictions from sunrise and sunset. The solar position will
be calculated using the centroid of the layer used.
}
\examples{
\dontrun{
library(landtools)
layer <- casas01
dayfootprint(layer, day = "2021-12-21", H = "Altura", outdir = "./outputs/", 
tz = "Europe/Paris", Xmin = 30, stylefile = "./styles/footprint_azul.qml")
}

}
