% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayintervals.R
\name{dayintervals}
\alias{dayintervals}
\title{dayintervals}
\usage{
dayintervals(
  location_geo,
  day = "2021-12-21",
  tz = "Europe/Paris",
  Xmin = 30,
  sunrisesunset = TRUE
)
}
\arguments{
\item{sunrisesunset}{if TRUE --> include sunrise and sunset in the final vector}

\item{t1}{time to iniciate --> format --> "2021-12-21 05:00:00"}

\item{t2}{time to stop --> forma --> "2021-12-21 05:00:00"}

\item{s1}{sunrise (must be calculated from geographic coords)}

\item{s2}{sunset (must be calculated from geographic coords)}

\item{cadaXmin}{number of minutes between each calculus}
}
\description{
This functions creates a list of date times for a whole day
between sunrise and sunset
}
\examples{
\dontrun{
layer <- casas01
location_geo <- locationgeo(layer)
dayintervals(location_geo, day='2021-08-24', Xmin=30)
}

}
