% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dftopoints.R
\name{dftopoints}
\alias{dftopoints}
\title{dftopoints}
\usage{
dftopoints(
  df,
  epsg1 = "EPSG:4326",
  epsg2 = "EPSG:25829",
  AtriCoords = c("lon", "lat")
)
}
\arguments{
\item{df}{data.frame with coords in attributes}

\item{epsg1}{is the epsg from origin in format "EPSG:00000" ("EPSG:4326" by default)}

\item{epsg2}{the epsg fro transformation if needed, in format "EPSG:00000" ("EPSG:25829" by default)}

\item{AtriCoords}{vector with the names of the attributes with "lon" and "lat" values (in that order)}
}
\description{
get a data.frame with lon lat and creates an spatial object converting CRS
}
\author{

}
