% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir2ovr.R
\name{dir2ovr}
\alias{dir2ovr}
\title{dir2ovr}
\usage{
dir2ovr(dir, pattern, parallel = TRUE, ncores = 12)
}
\arguments{
\item{x}{adsfasdf}
}
\value{
VRT
}
\description{
this is a function
}
\examples{
\dontrun{
# execute without parallel
dir2ovr("./data/raster/MDT25/", "*.asc", FALSE)
# execute parallel
dir2ovr("./data/raster/MDT25/", "*.asc", TRUE, 12)
}
}
