% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodir.R
\name{dodir}
\alias{dodir}
\title{dodir}
\usage{
dodir(path, removeall = FALSE)
}
\arguments{
\item{path}{}
}
\value{
string
}
\description{
This function sets a dir name and create the folder if doesn't exist
The penultimate folder must exist ¡¡
}
\examples{
\dontrun{
dodir("./ei/ou/")
}
}
