% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpath.R
\name{fpath}
\alias{fpath}
\title{fpath}
\usage{
fpath(path)
}
\arguments{
\item{x}{adsfasdf}
}
\value{
string (dir)
}
\description{
get the folder of a file or folderpath
}
\examples{
\dontrun{
fpath("./path/file.tif")
}
}
