% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationgeo.R
\name{locationgeo}
\alias{locationgeo}
\title{locationgeo}
\usage{
locationgeo(layer)
}
\arguments{
\item{layer}{geographic layer with the obstacles to calculate footprint}
}
\description{
creates a single points layer with the centroid in geographic coords
usable by shadow package as location for sun position
}
\examples{
\dontrun{
layer <- casas01
mapview(locationgeo(layer), color='red')+layer
}

}
