% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir2prj.R
\name{prj2dir}
\alias{prj2dir}
\title{prj2dir}
\usage{
prj2dir(prjfile, dir, pattern)
}
\arguments{
\item{x}{adsfasdf}
}
\value{
VRT
}
\description{
add a .prj file to each of the rasters within a dir 
All the rasters must have the same projection
}
\examples{
\dontrun{
dir2vrt("./data/raster/MDT25/", "./data/raster/", "MDT25.vrt", "*.asc")
vrt <- paste0("./data/raster/MDT25/", "MDT25.vrt")
mapview(vrt)
}
}
