% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplefootprint.R
\name{simplefootprint}
\alias{simplefootprint}
\title{simplefootprint}
\usage{
simplefootprint(
  layer,
  heightattr = "Altura",
  outdir = "./outputs/",
  time = "2021-12-21 12:00:00",
  tz = "Europe/Paris"
)
}
\arguments{
\item{layer}{geographic layer with the obstacles to calculate footprint}

\item{outdir}{output folder. A subfolder will be created}

\item{tz}{time zone ("Europe/Paris" by default)}

\item{heighattr}{attribute of the layer that contains the height of the obstacles}

\item{time1}{time to start calculation}

\item{time2}{time to stop calculation}

\item{stylefile}{path to the style file asociated to the results (not mandatory)}

\item{cadaXmin}{number of minutes between each calculation}
}
\description{
This function craetes a single gpkg with the footprints of 
the sun over a layer of buildings (with heigth attribute). The footprints will
be calculated from a vector of date-times, created internally. This vector is
based in the start time (t1) and stop time (t2) and takes into consideration
the possible time restrictions from sunrise and sunset. The solar position will
be calculated using the centroid of the layer used.
}
\examples{
\dontrun{
layer <- casas01
x <- simplefootprint(layer, time = "2021-12-21 10:00:00")
mapview(x)
}

}
